zero warning on build

+++ LudumGameEngine/Src/Components/CameraShakeComponent.h
@@ -84,9 +84,9 @@ public:
 			float shakeFreqX = 10;  float shakeFreqY = 8;
 			float shakeFreqY2 = 20;  float shakeSizeX = 15;
 			float shakeSizeY = 10;  float shakeSizeY2 = 5;
-			float t = SDL_GetTicks();
-			int xAdjustment = (int)sin(t * shakeRateX) * shakeSizeX;
-			int yAdjustment = (int)sin(t * shakeRateY) * shakeSizeY + cos(t * shakeFreqY2) * shakeSizeY2;
+			float t = (float)SDL_GetTicks();
+			int xAdjustment = (int)(sin(t * shakeRateX) * shakeSizeX);
+			int yAdjustment = (int)(sin(t * shakeRateY) * shakeSizeY + cos(t * shakeFreqY2) * shakeSizeY2);
 			Game::camera.x += xAdjustment;
 			Game::camera.y += yAdjustment;
 			cameraShakingUpdatesLeft--;
